Report inactive host controllers
It is interesting to know if host controllers are not active. I assume
that those inactive controllers are not using System Management Mode.

The last few days have been quite productive and I basically completed
the goal of having "USB Legacy Support" disabled. There might be some
devices that do not behave according to specifications, and there are
many nontrivial things that can go wrong. Anyway, this is a milestone.