Fix interrupt requests
The order of enabling IRQ 0 and IRQ 2 is relevant. Usually there is an
an interrupt source override (from 0 to 2) in the ACPI tables. I have
implemented a solution for cases where that is missing and the timer
seems to be using IRQ 2. However, I just noticed that the special code
is run even in cases where the override has been defined. It worked
almost all the time "by accident", but not like it was intended.

The fix is to enable IRQ 2 first and then IRQ 0. It will then override
the correct entry on the I/O APIC's Redirection Table.