Change usb_init_early behavior
The BIOS boot loader does not support memory addresses above 4 GiB. It
is not a good idea to return an error value because the early function
can be skipped. The log file has information but is not essential, and
when using the UEFI boot loader it would not be available anyway.

This applies to a very unlikely situation, i.e. the computer uses BIOS
and has PCI base addresses mapped above 4 GiB. I assume that this kind
of setup would have problems with other operating systems.