Add new fonts
These are "Bitstream Vera" fonts that have been converted to a format
that is compatible with Dancy. Minor modifications were needed. The
license required that fonts have to be renamed so I gave new names:

        Dancy Sans
        Dancy Sans Bold
        Dancy Sans Mono

The license is embedded in the file format. All changes made by me are
in public domain.