Move early boot code
The following files contain boot records:

        floppy.c - Volume Boot Record for FAT12/FAT16 floppies
        vbrchs.c - Volume Boot Record for FAT12/FAT16 (CHS)
        vbrlba.c - Volume Boot Record for FAT12/FAT16 (LBA)

All of them will load a second stage loader:

        ldr512.c - Second Stage Loader for FAT12/FAT16

In 2015, I was interested in learning the machine language and I used
a lot of time to compose the binary blobs. Debuggers and disassemblers
were the tools I used, and there are no proper source code files. All
instructions are intended to be compatible with "CPU 8086".