Modify release.mk
One "1.44M" floppy or two "720K" floppies can contain more data. At
least VirtualBox 6.1 supports booting fdd720a.img. The loader said:

         INSERT NEXT BOOT DISK
        PRESS ENTER TO CONTINUE

I used "Devices -> Floppy Drives -> Choose a disk file..." and chose
the "B" disk. An interesting experiment, but perhaps not very useful.
The main point of having the "720K" floppy is to test the loader with
different file system settings.

The loader itself supports different sector sizes, and I once tried
to boot a 1024-byte-sector floppy on real hardware. The BIOS did not
recognize it as a bootable disk. However, it could be possible to use
that as a second disk because then the loader is already in charge.