Add kernel.asm
The kernel_jump function is not very complicated. However, the caller
must allocate the stack and have it correctly set up, according to the
calling conventions. Like the name says, this "jumps" to the function
and this detail is important for the stack pointer alignment.

Because this function does not return, all general purpose registers
can be modified and there is no need to save any previous values.