Develop control transfer
This was more complicated than what I thought. I still don't have the
mechanism for transferring data structures from "init" to "kernel". If
building the ACPICA module, the linking stage will result in an error
message because the OS layer functions have not been implemented.

A page fault will occur when the kernel start functions return. This
behavior is for testing the error handling mechanism. Those functions
must not return when implemented because the return address is NULL.