Add apic.c
The apic_id function can be called even if APIC was not enabled. All
CPUs supported should have the Local APICs but the init module does
not necessarily enable them. In this case, enabling means configuring
the Spurious-Interrupt Vector Register(s). There is a bit that tells
whether the APIC is "software enabled" or not.