Update copyright years
I used some time to research can I use a range instead of individual
years. What I found out is that the range can be used if there really
were copyrightable changes in every year. I will still use individual
years in the license headers if the list does not become too long.

From now on, I'm going to pay more attention to whether the changes
really are copyrightable or not. Changing the year itself in the file
would not make that particular file changed in a "copyrightable" way.

This README file is special because it refers to the whole project.