Update memory_disable_manager function
This function is a bit like EFI_BOOT_SERVICES.ExitBootServices(), but
it is applied to the pre-kernel environment. The simple memory manager
does not own the memory map after the memory_disable_manager function
has been called.

All memory ranges that can be used like "EfiConventionalMemory" are
reserved for the kernel. This commit changed it to include databases
also. I think that those files should be part of the init module only.

Although the "simple" memory manager is not very simple anymore, and
could probably be simplified by rewriting it from scratch, the memory
map that is exported to the kernel is very simple and elegant.