Export memory map
The memory map is very essential. The system firmware (BIOS or UEFI)
is the original source of the information, but the memory map has been
processed by boot loaders and the pre-kernel environment.

The "kernel->memory_map_size" member does not provide any information
that could not be calculated by examining the map itself.