Verify memory map
I noticed that something was wrong because the memory_entries variable
value was not correct after calling the b_exit() function. This was
only when using the UEFI boot loader. After some debugging, the reason
was very clear and the boot loader worked just like it should. I did
not take into account that my UEFI boot loader modified the memory map
and the previously calculated value was not correct anymore.

Although it is not needed to call this memory_init function again if
using the BIOS boot loader, I do it anyway because it checks that the
memory map is still valid. It should reveal possible bugs and in some
unlikely cases it could even detect hardware faults if the memory map
did not contain previously written values.