Change data types
It is very common to define "int i" and use it in for loops. Although
the "count" values must be positive, I'm going to use signed integers.
In theory, a compiler can optimize more because signed overflow is not
allowed but I have not checked how relevant that is in practice.