Initialize kernel memory
This makes the start-up procedure a little bit slower. However, I did
not like that there were undefined values in the kernel memory block.
It would be better if they were truly random values but most likely
the undefined values are zero bytes. There could be bugs that do not
show up until there are other values.

When allocating and freeing kernel heap memory, the content will be
"undefined". At the beginning, the state is deterministic.