Print date and time
The reason for this is to guarantee that something is printed before
starting the kernel. Otherwise there would be no window where to print
if using the "kernel->print" function pointer. At that time, it is not
possible to create a window because the init memory manager has been
disabled.