Ignore ident directives
There were unnecessary ".rdata$zzz" sections that contained the same
data many times. I noticed this when I wrote my linker but I left it
like that because it was a good test to see that sorting the sections
worked like it should. In the final object, it was likely that those
"zzz" sections were placed at the end of the ".rdata" section.