Introduce Global Descriptor Table
The GDT has a long history. The Dancy Operating System and most of the
other systems use only the features that are really needed. The memory
model is "flat" so that segmentation is effectively turned off. There
are two exceptions: segments FS and GS may have base addresses other
than zero for processor or thread information blocks.