Develop table.h
The section start addresses and sizes are added in the kernel table. I
have not decided should I use large page sizes in the kernel space. If
yes, then the section information is not very useful because .text and
and .rdata sections would not be marked as read-only.

The global symbol table is useful for many purposes, e.g. for finding
the nearest symbol name of an arbitrary address.