Allow nondeterministic processor order
I noticed that at least in VirtualBox the warning message "unexpected
execution order" was printed every now and then. I removed the whole
message and treat it like if the order of application processor kernel
jumps were nondeterministic. The interprocessor interrupts are sent in
a defined order so it is still likely that the order is deterministic.

However, now that one extra delay is removed, the "random" order will
happen often enough (because it seemed to happen even with the delay)
so that there should not be hidden dependencies on a specific order.