Remove init module threads
The gui_print function will refresh the screen. I removed threads from
the init module because they are not really needed. It was a very good
exercise to try them and I did not find any problems or bugs. It would
require more testing, e.g. many threads created and terminated, to be
able to say that they are stable. Because that would be artificial in
the init module context, it is better to have a simpler implementation
and leave the threads to the kernel.

Each processor has one thread of execution in the init module.