Remove gui_move_window function
This function has not been used and I am not going to develop the init
module in a direction that would need it. There will be a proper GUI,
with movable windows, but it is implemented in user space. However, it
will take some time until I have enough infrastructure for that.

Also, I made a very small change in the gui_print function. The blit
function is called directly while the mutex has already been acquired.

Today I have done many changes that have simplified the init module
and the Dancy source tree in general. It has been said that throwing
code away can be very productive.