Add idt_install_asm function
There will be cases when it is reasonable to have an interrupt handler
that has been optimized for specific interrupts. The kernel timer is
a very good example because that handler code will be heavily used. It
would have been possible to replace the previous handler atomically if
using DANCY_64, but I wanted to keep it consistent with DANCY_32.

The caller must take into account that there is a short time when the
interrupt descriptor is not present (flag P is 0). If the present flag
is set, the descriptor is valid and not partially updated.