Introduce timer handler
The timer interrupt has a special handler. The timer_handler function,
which is written in C, is called when the timer_ticks_wait variable is
zero. This interface should make it possible to have a "tickless" mode
implemented in the future. Now I am simply more interested in having a
fast handler for cases when the interrupt skips the C function.

The DANCY_32 implementation is a bit different because it uses the ebp
register in the early handler. The fact that user space code can load
certain segment descriptors is the reason for many workarounds. Also,
I want that the kernel does not change them if the user space code for
some reason does that (e.g. mov eax, 0; mov ds, eax).