Map individual pages
I will try to use the "standard framebuffer" in a way that only the
modified pages will be written to the "physical" framebuffer. This is
the reason I wanted to guarantee that large pages are not used for
mapping the fb_standard_addr buffer.

In general, large pages provide advantages when it comes to the "TLB"
(translation lookaside buffer), but at this point I have not really
thought about that aspect. It will become more relevant after I have
more kernel functionality implemented and processes running in user
space.