Develop heap.c
I added a new "heap_alloc_static_page" function. There is no mechanism
for freeing the returned page, and that is why it is called "static".
The returned page is initialized to zero because I think this is very
consistent with the C environment as far as "static" is concerned.

When I implemented the function mentioned above, I noticed that there
was a bug in the existing code. If increasing the heap map size, I did
not test correctly whether there was enough free space to do that.