Add pg_get_entry function
This function returns a pointer to the page-table entry (PTE), which
points to the physical page. Only 4 KiB pages are supported by this
function and it will return a NULL pointer if 2 MiB or 4 MiB is used
for the input address (const void pointer, actually). I implemented
it like this because this avoids doing the page size test later, and
I assume that I will use this only for 4 KiB pages.

The idea is to provide a mechanism for accessing the Dirty (D) bit.
I will create some kind of array of pointers that describe the whole
"standard framebuffer" and render only the modified pages into the
real framebuffer. Details are not clear yet.

Also, there may be other important uses for this function.