Add fb.c
I really like this code. There still is potential for optimizations,
but now the "algorithms" are efficient. This description should not be
taken literally. I mean that copying the modified framebuffer areas
instead of everything is a very powerful technique. Precalculating the
framebuffer destination offsets and sizes for each page could be worth
mentioning too. This was an easy way to handle the last page, which
may have unused bytes (e.g. 800x600 is not page-aligned).