Detach init module
I added a new function to the kernel table. The basic idea is the same
as in the EFI_BOOT_SERVICES.ExitBootServices() function. After calling
the new function, the kernel becomes responsible for everything. All
the memory areas that were used by the init module can be reused.

The standard framebuffer will contain the screen content.