Render kernel console glyphs
The init module uses my TrueType renderer (common/ttf.c) and provides
the kernel with pre-rendered glyphs. I will not implement any kind of
graphical interface in the kernel, but some kind of kernel console is
needed. For example, the panic() function needs a mechanism to print
something on the screen. While developing the operating system, it is
very useful to have "printf debugging" tools available.