Refactor standard memory functions
Compilers can do some optimizations, e.g. "malloc + memset" is handled
by "calloc", and this would have resulted in an infinite recursion. I
have not checked all the details, and perhaps it would be possible to
prevent it somehow. However, the most robust solution is to avoid that
risky coding convention (calling "malloc" when implementing "calloc").