Optimize console.c
This optimization is very important, and not having it would be like a
bug. The whole idea is to avoid rewriting the same data again into the
framebuffer. In practice, there are so many empty spaces in the buffer
all the time and those must not be rewritten when scrolling.