Use new fb_copy_vga function
I cannot find the source of this information, but I once read that
there might be problems if doing "rep movsq" on the VGA framebuffer.
I'm not sure whether this is true or not. However, I use "rep movsd"
also on x64 just to be sure. It is very likely that 64-bit computers
support VBE (BIOS) or GOP (UEFI) video modes.