Update detach_init_module function
The init module shows a window but it would disappear too fast without
this "sleeping" feature. It is possible to remove this later. I wanted
to put the logic in the init module because it is a bit controversial.
If there were more time-consuming procedures before calling the detach
function, this feature itself would not slow down the start-up.