Fix scrolling
Now the console should work in a logical way if setting the scrolling
region and then changing the cursor position so that it would not be
inside that region. As far as I know, the exact behavior has not been
defined very well. Maybe this is implementation-defined behavior, and
my version chose to behave this way.