Change line wrapping behavior
This feature is quite complicated but hopefully I took all relevant
corner cases into account. In this case the word "corner case" can be
interpreted literally. Now it is possible to write a character in the
bottom right corner without scrolling the screen.

Also, if writing a full row of characters, the next carriage return
would still work on the same row. In the same circumstances, the line
feed would not give the (most probably) unwanted extra blank line.