Implement visible cursor
The cursor is rendered so that alpha blending is used. It is possible
to see the glyph that is under the cursor. Even though the "glyph" is
usually an empty space, exactly the same technique is applied. I was
wondering should I have a visible cursor by default. Now the default
setting is to keep it invisible but this can be changed easily.