Improve console.c
In the previous commit I added code for decoding UTF-8 sequences and
now that code is used. When doing tests I noticed that my "con_print"
function had a bug and it basically did not release the lock if using
it like this: con_print("%s", "");

For an ill-formed UTF-8 sequence "\xF0\x80\x80\x41", an example from
the Unicode Standard, my console will print codes U+FFFD and U+0041.