Add task.h
This is a very fundamental part of any kernel. I need to study more on
the topic, especially the symmetric multiprocessing (SMP) aspect. The
most important thing is to get started. Most likely the "struct task"
will be one of the most important data structures. At least something
that will be looked at if evaluating the kernel behavior.