Develop task management
There are many things that need to be taken into account. My initial
implementation is not complete but so far it seems to be a good start.
Maximum performance is not the primary goal at this point. That will
be an important aspect later when I have good benchmarks and relevant
performance bottlenecks have been reliably identified.