Develop task management
I added one "abstraction level" to the task_switch function. There can
be many CPUs involved and I wanted to make sure that the same task was
not started up in parallel. It is safe to call "task_switch" function
for the currently running task, e.g. "task_switch(task_current())". It
would not call the assembly procedure because the flag tells that the
task was already running.