Remove unnecessary instructions
In the init module, I had an idt_handler function which took the error
code as a parameter. Then I removed that because the stack parameter
can be used for reading the code. However, I forgot to remove one line
of assembly that was related to this previous feature.

As a small detail, it would be possible to use the "fastcall" calling
convention on IA-32 if first parameters were in registers ecx and edx.
In this idt_handler function that is the case but I don't remember was
that really intentional or did I just use the same "assembly style" as
in the x86-64 calling convention.