Refine task identity
There was a very big problem I could not get over. If I save the task
structure pointer somewhere, it is hard to guarantee that it would not
be used for dereferencing a "dangling" structure later.

Some time ago I decided that a "task id" in Dancy Operating System is
uint64_t and the values are not reused (until rebooting the OS).