Add sti instruction
All external maskable interrupts are always enabled at the end of the
task_switch_asm function. The caller is responsible for making sure
that this behavior is safe before calling the function. It is noticed
that "sti and ret" will allow "ret" to be executed without enabling
the interrupts if they were disabled originally. However, nothing is
strictly relying on this behavior but it is still worth mentioning.