Use task_yield function
This is a very simple implementation. The notion of fairness in mutex
acquisition has not been taken into account. When I'm more experienced
and have done enough benchmark tests, I will get back to this topic,
if necessary. It _could_ be possible that this is not a bottleneck if
other parts were working optimally enough.