Develop task management
There are cases when task switching should not occur. This new pair of
functions, task_switch_disable and task_switch_enable, can be used for
solving the issue. The callers have a great responsibility for their
proper use, but I hope that there would not be too many of them. These
functions are nestable and I was thinking should I allow calling the
enable function even if task switching were enabled already. However,
most probably it would be a bug so I decided that it is not allowed.