Improve task.c
It is important to disable interrupts before calling the gdt_get_tss
and task_switch_asm functions. If a timer interrupt had changed the
task after getting the TSS, it would possibly have been pointing to
a wrong structure later when switching back to the same task. Though
on multiprocessor systems only.