Add task_exit function
The thread of execution is still using the kernel stack of this task.
It will have the "active" flag set until task switching happens. There
is not much to do at the moment but this function might have a little
bit more logic in it when the operating system is more advanced.