Improve handler of exception vector 13
This applies only to kernel code. This special handler will try to set
segment registers es and ds to the same value as the segment register
ss before calling the actual handler. This is relevant if a user space
thread modifies the segment registers in an allowed way, e.g. setting
them to zero, and this makes kernel code trigger a general-protection
exception when those same segments are used.

The zero segment case was already handled before this commit. However,
as a more robust solution I wanted to make sure that the values are
exactly the same as in the "known-to-be-correct" stack segment. This
so called pre-handler might be run two times without fixing the real
reason for the exception. It should not matter because efficiency is
not the primary concern if kernel code triggers an exception.