Improve idt_panic function
One of the basic features of an error screen is to print some context
information. For normal computer users, that information is not very
helpful and it would be important to avoid such exceptions. Perhaps
the only valid reason would be hardware failures. In practice, there
is a software bug in the system if the error screen is shown.