Add mm.c
This is the first version of my physical memory manager. I was worried
about its performance, but it was better than what I expected. I can
improve this later when other parts of the operating system work and
I have good data to show where the real bottlenecks are. If this part
had a minor effect, I would probably appreciate its simplicity more.