Develop user space paging
It is possible to use the pg_map_user function many times for the same
virtual memory area. All user space pages of the task are free'd when
deleting the virtual address space. It might be that some pages need
to be preserved if they are shared in the future. There are available
bits ("AVL") which could be used for marking such special mappings.

The latter was one reason I made the static pg_free_pte function. Now
it is very simple and could have been inlined into the for loop. There
would have been too many tab indents for the 64-bit version, however.