Limit task structure count
I did not want that all physical memory could be allocated for task
structures. Once a task structure is created, it is not freed but may
be reused later. If there is a pointer to a task structure, and that
could be saved in many places later, it can always be safely given to
the task_switch function. It does not necessarily refer to the same
task but the structure itself is valid.