Develop irq.c
I added functions irq_disable and irq_enable. Performance has not been
optimized for PIC (e.g. caching the masks) because I wanted to keep it
simple and reliable. The best optimization is to avoid calling these
new functions too often. This is especially important for the I/O APIC
versions because those might be a little bit slow.

I have not actually checked the performance. It might be that the port
inputs and outputs are slower than the seemingly complicated logic in
the I/O APIC functions.