Develop irq.c
I added functions irq_install and irq_uninstall. The same IRQ can be
shared so that there could be many handlers.

There is one coding convention that is a little bit unclear: I assume
that pointers are modified in an atomic way. The C standard definitely
does not guarantee that, but on the supported platforms it would sound
like a very reasonable "implementation-defined" behavior. I will need
to study this topic (e.g. stdatomic.h in C11).