Add memory access functions
Functions pg_read_memory and pg_write_memory can be used for accessing
any physical memory address. Supported sizes are 1, 2, 4, and 8 bytes.
It is meant that the caller uses naturally aligned addresses, but this
version allows unaligned accesses if page boundaries are not crossed.

It is not a coincidence that interfaces are quite similar with ACPICA
functions AcpiOsReadMemory and AcpiOsWriteMemory.