Implement PCI configuration space access
Functions AcpiOsReadPciConfiguration and AcpiOsWritePciConfiguration
are complicated. I can not guarantee correctness. It is known that
writes are not atomic if using other widths than 32-bit, but I assume
that the caller does not rely on that.

My current native interface works on 32-bit reads and writes because
it sounded like a good idea. Those 8, 16, and 64 bit versions should
not be needed in my own code, but time will tell.