Improve task_exit function
There is a task created for each CPU that is in use, but those are not
actually specific to the particular CPUs. The main point is that there
are as many tasks (or more) as there are active CPUs.

The above means that those "system" tasks shall not be exited.