Develop runlevel_reset function
The "triple fault" method for rebooting the computer is not meant to
be used. Instead of that, older computers should support and use the
"8042" method. However, the panic message is still there:

        con_panic("It is safe to reset the computer.");

It should be very likely that the computer is somehow rebooted when
calling the runlevel_set(6) function. Like mentioned in the previous
commit message, the AcpiReset function (ACPICA) is the primary way.
It should be a robust and "professional" method for doing the reset.