Improve initialization messages
I decided to print more accurate information about the status of the
USB host controllers. Also, it is more elegant to move these b_print()
function calls, that do not indicate errors, out from the usb.c file.

As a general note, it does not look very aesthetically pleasing when
using a non-monospace font and trying to print messages like these:

  Enhanced Host Controller Interface (EHCI): OK
  Extensible Host Controller Interface (xHCI): OK

Even though OK columns were not aligned with a monospace font either,
the look and feel is somehow worse with a proportional font.