Fix ps2_init function
There was a bug in this "configuration byte write" routine. The status
bit for the input buffer must be clear before writing the byte. What
makes this interesting is that the buggy routine actually failed on my
test laptop. I assumed that a ps2_delay() call would have given enough
time for it to work even with the erroneous logic.

Anyway, it was very good that it failed so I found the bug.