Simplify AcpiOsSleep function
The interface is very similar and it was not a coincidence. The sleep
function implementation has been moved out from the "Services Layer
for ACPICA" code.

I have not done many experiments with ACPICA, but the features I have
tried seem to work very well. This integration was a very good idea
for many reasons. First of all, it was a very good guideline for the
project to implement the services layer. Secondly, I honestly think
that some ACPICA features will be very useful later.