Add pci_init function
Although this function does nothing at the moment, I think it would be
very consistent to have it.

I really like the new "structure" of the kernel. It did not feel right
to put too many features in the "base" module, e.g. "PCI" or "PS/2". I
assume that from now on the kernel source code grows quite naturally.