Develop mutex functions
Whenever trying to lock the mutex, the interrupts cannot be disabled.
I assume that this strict requirement is a good thing in the long run,
but I am not absolutely sure. All "mtx_lock" and "mtx_trylock" callers
must check the return value and handle "thrd_error" in a defined way.