Add event.c
I have been reading about events and other synchronization primitives
and it seems that there is no universal definition for an "event".

One thing worth mentioning is that my event_wait function always uses
the timeout, which is uint16_t in milliseconds. This will guarantee
that callers must handle the return value and perhaps repeat the call
again in a for/while loop. Having longer timeouts may result in bugs
that are simply hidden until timeouts expire in the distant future.

The rationale above tries to follow the principle of having things
happening in a relatively compact time scale. For example, if running
the system for an hour, it should have covered as many "corner cases"
as possible. As far as testing is concerned, it is very difficult if
there were long timeouts that might wake up sleeping bugs much later.