Fix compiler warnings
Two warnings (-Wint-to-pointer-cast and -Wpointer-to-int-cast). This
does not happen too often nowadays, and I agree that these warnings
were technically valid. Maybe the second warning for

        uint64_t data0 = (uint64_t)event;

should be questioned. It is true that on 32-bit systems it is a "cast
from pointer to integer of different size". However, it does not lose
information because sizeof(void *) < sizeof(uint64_t).

Instead of arguing, let's be nice for the compiler and fix the code.
Usually the compiler knows the best, and rare "false positives" are
totally acceptable. They were technically valid, to say it again.