Use atomic pointer modification
The type of the "next" member has been changed. It is an "addr_t" and
all accesses are handled by specific functions. This has bothered me
for a long time. I know that it is very likely that C compilers would
have used atomic read and writes for pointers on supported execution
environments.

However, my design strongly relies on the fact that "next" pointers
are either valid pointers to task structures or NULLs. Also, there
were "compilers could reorder statements" concerns. Now there must be
a sequence point whenever calling an external function.

By the way, it was very easy to do this conversion. First I changed
the type in the task structure and then I tried to compile the source
code. The compiler warnings told me what I needed to fix.