Add task_find function
This function returns a task structure that has, or had at some point,
the specific identification. These IDs are not reused.

To avoid possible deadlocks, it is a strict requirement that interrupt
contexts cannot call this function. First I thought that maybe a NULL
should be returned, but that would potentially hide bugs. I ended up
having a system panic message so that the rule cannot be ignored.

In general, keeping interrupts enabled is very crucial for the system.