Add task event functions
The "write" function will set up the event function and argument data
in an atomic way. Disabling interrupts is enough because the same task
cannot be active on any other CPU at the same time.

The "read" function returns the value from the event function. If it
is zero, the function will be replaced by a default event function,
which always returns zero. The "read" function can always be called as
many times as wanted, without never setting the custom event function.