Add new task_switch function
This version contains a lot of comments. I have been thinking about
this topic a lot for the last few days. Assuming that there are no
implementation bugs, the overall design should be correct. I will do
a lot of tests for both correctness and performance. The former has
been the most important aspect, and it can be measured objectively.