Add task_list function
This will enumerate all tasks that are in the circular linked list. It
is caller's responsibility to allocate the buffer. This is an example,
without proper error checking:

        struct task_list_entry *buf;
        size_t buf_size = sizeof(*buf) * BUFFER_ENTRIES;
        int i, count;

        buf = malloc(buf_size);
        count = task_list(buf, buf_size);

        if (count > BUFFER_ENTRIES) {
                /*
                 * The buffer does not contain everything, but
                 * is valid otherwise.
                 */
                count = BUFFER_ENTRIES;
        }

        /*
         * Check all task identifications.
         */
        for (i = 0; i < count; i++) {
                uint64_t id = buf[i].id;

                /*
                 * Do something with the ID.
                 */
        }

        free(buf);

        /*
         * Get the number of tasks.
         */
        count = task_list(NULL, 0);

Perhaps it makes sense to first check the number of tasks, and then
allocate the buffer based on that. Because the number of tasks is not
stable, it is a good idea to add some extra space.