Improve mtx_lock function
This important optimization is almost a must-have. If the lock can not
be acquired, the "task event" function will be set. It is important to
notice that these task_write_event and task_read_event functions can
be called even before initializing the actual task management. If that
is the case, the task_yield function returns immediately and the mutex
works like a spinlock.