Develop 8042.c
The circular buffers have been implemented for PS/2 ports. The event
is "signaled" if there are data in the buffer. The signaled state is
automatically reset when all the data have been read from the buffer.

I have mentioned this several times in my commit messages: the init
module disables the "USB Legacy Support" so it is not possible to
use USB keyboards and mice until I have "real" USB code.