Redefine task hierarchy
The application processor tasks have owner ID 1. Only the first task
has owner ID 0, and it is the only task that is not accepted by the
task_exit function.

There is always a task for each processor, e.g. if the application
processor task exits, it will not release the current task structure
until another runnable task is found.