Improve fix_memory_map function
The memory map entries that have "length 0" must be removed. The boot
loaders did not give those kind of entries to the init module, but my
memory allocation functions could do that. If the base address of an
allocation were the same as the base of "free entry size", the memory
map would end up having a "free entry" with "length 0" plus an entry
for the allocation itself.

        void *p1 = malloc(16);
        void *p2 = malloc(16);
        void *p3 = malloc(16);

        free(p2), p2 = malloc(16);

The above code would have reproduced the bug.