Renew delay_calibrate function
I have had a lot of problems with Time Stamp Counters (TSC) on virtual
computers. If I remember correctly, and assuming there were no bugs in
my own code, the RDTSC instruction sometimes returned a value that was
less than the value read previously. On real hardware, and on the same
specific processor, that should not happen.

This new version of the delay_calibrate function has been simplified.
It will wait 100 milliseconds and then calculates the difference.