Verify memory availability
It is hard to guarantee anything if physical memory addresses do not
work correctly. However, the bitmap should be checked and everything
must be "reserved" initially. This is the safest assumption, and even
if changing from "reserved" to "free" failed, the behavior would be
at least somewhat defined.

This strategy should detect the case if the memory map (from the boot
loader) says that "range from 0x100000000 to something" is there, and
available for use, but those addresses returned zeros no matter what
is written to them.