Improve task_schedule_default function
This little cpu_halt(1) function call elegantly solves the challenge
I have been trying to figure out. This is like an "idle" task, but it
is implemented in the scheduler.

Checking the interrupt flag is important for two reasons. The first
one is for making sure that the halted CPU will be woken up. The other
reason is subtler. If the task is preempted by the timer handler, and
there are no other runnable tasks, the current task continues without
being halted for a moment. The timer handler has interrupts disabled.