Refactor runlevel.c
The function name "runlevel_set" has always looked awkward. Not just
that, but it can also be read like in the passive voice, although it
was meant to be a verb.

The new name "runlevel_send_request" has a deeper meaning. It really
is sending a request to another task and the longer function name may
be justified because of its strong functionality (e.g. shutdown).